requirements.txt
asgiref==3.6.0
certifi==2023.7.22
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
# Django>=3.2.4,<4.0.0
django-cors-headers==3.14.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
drf-yasg==1.21.5
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
openapi-codec==1.3.2
packaging==23.0
Pillow==10.0.1
pytz==2022.7.1
requests==2.28.2
rest-framework-simplejwt==0.0.1
ruamel.yaml==0.17.21
simplejson==3.18.3
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==1.26.14
gunicorn
dj_database_url
psycopg2
django
djangorestframework-simplejwt